refactor: migrate components button-switch-theme with its operating logic; remove unused elements
