+++ src/app/providers.tsx
@@ -3,7 +3,12 @@
 import { Provider } from 'react-redux';
 
 import { store } from './store/store';
+import { ThemeProvider } from './utils/context/context';
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  return <Provider store={store}>{children}</Provider>;
+  return (
+    <Provider store={store}>
+      <ThemeProvider>{children}</ThemeProvider>
+    </Provider>
+  );
 }
