+++ src/app/utils/context/use-theme.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { createContext, useContext } from 'react';
 
 interface IThemeContext {
